Remove all Go ebuilds.
The Go ebuilds are unmaintainable in their current form. Packaging
software written in Go in the traditional way - where the package
management system handles dependencies - is just too much hassle.

The syncthing-inotify ebuild is available in the go-overlay.

If you were using the Decred ebuilds and want them back, feel free to
open an issue. I have working ebuilds for dcrd-dcrctl and dcrwallet, but
they need to be cleaned up. If nobody uses them, it's not worth the
effort.