net-im/dendrite: make ebuild fail on errors
After a musl upgrade, parts of dendrite fail to build:

sqlite3-binding.c:37644:42: error: 'pread64' undeclared here (not in a function); did you mean 'pread'?

This results in the ebuild being installed without the actual dendrite
binary.

Make the ebuild fail if compilation fails, and while at it, also when
the modifications to the sample config fail.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>